requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Flask-Moment==1.0.5
Flask-WTF==1.1.1
Flask-Markdown==0.3
Flask-Bootstrap==3.3.7.1
pyotp==2.8.0
PyJWT==2.6.0
pycryptodome==3.17
PyQRCode==1.2.1
python-dateutil==2.8.2
requests==2.28.2
azure-identity==1.12.0
azure-keyvault-secrets==4.6.0
azure-keyvault-certificates==4.6.0
azure-keyvault-keys==4.7.0
marshmallow==3.19.0
Authlib==1.2.0
mysql-connector-python==8.0.32
mysqlclient==2.1.1
PyYAML==6.0.0
Flask-LDAP3-Login==0.9.18
xhtml2pdf==0.2.11